added flags wall werror wextra and fixed werror bugs

+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: rafaelfe <rafaelfe@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/19 11:44:14 by rafaelfe          #+#    #+#             */
-/*   Updated: 2025/03/29 15:42:41 by rafaelfe         ###   ########.fr       */
+/*   Updated: 2025/03/30 15:39:33 by rafaelfe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,7 +88,7 @@ t_token	*ft_tokennew(char *str, int type);
 void	ft_token_addback(t_token **token, t_token *new_token);
 int		is_space(char prompt);
 int		is_operator(char prompt);
-char	*ft_insertstr(char	*string, int index, char *substr);
+char	*ft_insertstr(char	*string, size_t index, char *substr);
 void	expand_tokens(t_token *token);
 void	create_cmds(t_shell *sh, t_token *token);
 void	execute(t_shell *sh);
